requirements.txt
streamlit>=1.31.0
altair>=5.0.0
pandas
fredapi
plotly
python-dotenv
google-generativeai
numpy
wbgapi
pytest
pytest-cov